+++ __tests__/__integration__/imperative/__tests__/__integration__/cli/auth/__scripts__/base_profile_and_auth_login_config_global.sh
@@ -12,14 +12,14 @@ then
     exit $CMDRC
 fi
 
-imperative-test-cli config set "profiles.my_base.properties.user" "$baseUser" --global-config
+imperative-test-cli config set "profiles.base.properties.user" "$baseUser" --global-config
 if [ $CMDRC -gt 0 ]
 then
     echo "Setting user of config failed!" 1>&2
     exit $CMDRC
 fi
 
-imperative-test-cli config set "profiles.my_base.properties.password" "$basePass" --global-config
+imperative-test-cli config set "profiles.base.properties.password" "$basePass" --global-config
 if [ $CMDRC -gt 0 ]
 then
     echo "Setting password of config failed!" 1>&2